requirements.txt
click==7.1.2
fastapi==0.65.1
h11==0.12.0
httptools==0.1.2
pydantic==1.8.2
pymongo==3.11.4
python-dotenv==0.17.1
pyyaml==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
